refactor(cli): extract logging setup into cli.log
The logger, the sd-daemon priority formatter, the TTY probe, the
--verbosity option, and the handler wiring had grown past what belongs
inline in main.py. Move them to cli/log.py, which owns how the
zfs.replicate logger's output is presented; main.py just calls
configure() and decorates with verbosity_option. The library modules
still emit through logging.getLogger(__name__) and are unaffected.

Pure refactor -- no behaviour change. The formatter tests move to
cli_test/log_test.py alongside it.
