+++ zfs/replicate/cli/main.py
@@ -1,11 +1,8 @@
 """Main function zfs-replicate."""
 
 import itertools
-import logging
-import sys
 
 import click
-import click_log
 
 from .. import filesystem, receive, send, snapshot, ssh, task
 from ..compress import Compression
@@ -14,50 +11,13 @@ from ..filesystem import filesystem as filesystem_t
 from ..ssh import Cipher
 from . import options
 from .click import EnumChoice
+from .log import configure, logger, verbosity_option
 
-logger = logging.getLogger("zfs.replicate")
-
-# Python level -> sd-daemon(3) / syslog priority.
-_PRIORITIES = {
-    logging.CRITICAL: 2,
-    logging.ERROR: 3,
-    logging.WARNING: 4,
-    logging.INFO: 6,
-    logging.DEBUG: 7,
-}
-
-
-def _stderr_is_tty() -> bool:
-    return sys.stderr.isatty()
-
-
-# click-log's ColorFormatter is untyped, so mypy sees an ``Any`` base class.
-class _Formatter(click_log.ColorFormatter):  # type: ignore[misc]
-    """Colored output on a terminal; ``<N>`` priority prefixes otherwise.
-
-    On a plain (non-terminal) stream -- most importantly the journal stream of
-    a systemd unit -- each line is prefixed with its sd-daemon priority so
-    journald records the level and ``journalctl -p`` can filter by it. On a
-    terminal the record keeps click-log's colored ``level:`` presentation.
-    """
-
-    def format(self, record: logging.LogRecord) -> str:
-        if _stderr_is_tty():
-            return super().format(record)  # type: ignore[no-any-return]
-
-        priority = _PRIORITIES.get(record.levelno, _PRIORITIES[logging.INFO])
-        return "\n".join(
-            f"<{priority}>{line}" for line in record.getMessage().splitlines()
-        )
-
-
-click_log.basic_config(logger)
-logger.handlers[0].setFormatter(_Formatter())
+configure()
 
 
 @click.command()  # type: ignore[misc]
-# Default WARNING keeps a plain run quiet; click-log's own default is INFO.
-@click_log.simple_verbosity_option(logger, default="WARNING")  # type: ignore[misc]
+@verbosity_option
 @click.option(  # type: ignore[misc]
     "--dry-run",
     is_flag=True,
